Move App Identity service to separate module
This makes it clearer where the actual service class resides.
